+++ slack-messaging/src/blocks/elements/builders.rs
@@ -23,4 +23,5 @@ pub use super::select_menus::{
 };
 pub use super::time_picker::TimePickerBuilder;
 pub use super::url_input::UrlInputBuilder;
+pub use super::url_source::UrlSourceBuilder;
 pub use super::workflow_button::WorkflowButtonBuilder;