+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.7.3][] - 2026-03-16
+
+- Support `Taskcard` block.
+- Support syntax-highlighted code blocks (language-specific).
+
 ## [0.7.2][] - 2025-12-18
 
 - Add missing builder exports in composition objects module.
@@ -123,6 +128,7 @@ And the `select menu element` and the `multi-select menu element` are renewed.
 
 - pre-release
 
+[0.7.3]: https://github.com/kaicoh/slack-messaging/releases/v0.7.3
 [0.7.2]: https://github.com/kaicoh/slack-messaging/releases/v0.7.2
 [0.7.1]: https://github.com/kaicoh/slack-messaging/releases/v0.7.1
 [0.7.0]: https://github.com/kaicoh/slack-messaging/releases/v0.7.0